Move component param logic out of manager.go
